♻️  Extract customerDataType
it will be needed for generating local storage key
