+++ packages/core/src/tools/serialisation/serializationConstants.ts
@@ -1,0 +1,6 @@
+export const enum CustomerDataType {
+  FeatureFlag,
+  User,
+  GlobalContext,
+  LoggerContext,
+}
